% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma.reports.R
\name{check.last}
\alias{check.last}
\title{Analyze the presences of commas (',') in the last column of a SQL database.}
\usage{
check.last(con, sql.tbl)
}
\arguments{
\item{con}{A `DBIConnection` object}

\item{sql.tbl}{The reference string of the table to be querried.}
}
\description{
Analyze the presences of commas (',') in the last column of a SQL database.
}
