% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.limited.class.encoder.R
\name{choose.categories}
\alias{choose.categories}
\title{Select a subset of a categorical variable cardinality.}
\usage{
choose.categories(values, cutoff = 0.8, min.use = 1, maxitems = NA)
}
\arguments{
\item{values}{Vector of values to reduce cardinality of}

\item{cutoff}{Total percentage of variable to capture}

\item{min.use}{Minimum use of each choice to be included}

\item{maxitems}{[optional] Maximum cardinality to reach}
}
\description{
Select a subset of a categorical variable cardinality.
}
