% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.data.R
\name{extract_name.POSIXct}
\alias{extract_name.POSIXct}
\title{Convert date time variables aggregated by `Groups` into a string representation.}
\usage{
extract_name.POSIXct(data)
}
\arguments{
\item{data}{The data set to be converted.
Should have columns `Value` and `Group`, where `Value` is a date time and `Group` is categorical.}
}
\description{
Convert date time variables aggregated by `Groups` into a string representation.
}
