% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.data.R
\name{extract_name}
\alias{extract_name}
\title{Convert numeric variables aggregated by `Groups` into a string representation.}
\usage{
extract_name(data)
}
\arguments{
\item{data}{The data set to be converted.
Should have columns `Value` and `Group`, where `Value` is numeric and `Group` is categorical.}
}
\description{
Convert numeric variables aggregated by `Groups` into a string representation.
}
