% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.limited.class.encoder.R
\name{make.limited.class.encoder}
\alias{make.limited.class.encoder}
\title{Produce a function to reduce carnality of feature vectors in a data frame.
Learns a pattern for each column by name.}
\usage{
make.limited.class.encoder(trainingData)
}
\arguments{
\item{trainingData}{A data frame to learn class encoders from}
}
\description{
Produce a function to reduce carnality of feature vectors in a data frame.
Learns a pattern for each column by name.
}
