% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.data.R
\name{prepare.data.default}
\alias{prepare.data.default}
\title{Method for `prepare.data` for any other data types (eg numeric vector)}
\usage{
\method{prepare.data}{default}(feature, target, extract_name = extract_name)
}
\arguments{
\item{feature}{Vector of data}

\item{target}{Vector of data where some values are missing}

\item{extract_name}{Function for naming discretized features}
}
\description{
Method for `prepare.data` for any other data types (eg numeric vector)
}
