% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.data.R
\name{prepare.data.POSIXct}
\alias{prepare.data.POSIXct}
\title{Method for `prepare.data` if feature is a date/time (POSIXct vector)}
\usage{
\method{prepare.data}{POSIXct}(feature, target, extract_name = extract_name)
}
\arguments{
\item{feature}{Vector of data}

\item{target}{Vector of data where some values are missing}

\item{extract_name}{Function for naming discretized features}
}
\description{
Method for `prepare.data` if feature is a date/time (POSIXct vector)
}
