% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.data.R
\name{prepare.data}
\alias{prepare.data}
\title{This is a generic method that dispatches based on the first argument.}
\usage{
prepare.data(feature, target, extract_name = extract_name)
}
\arguments{
\item{feature}{Vector of data}

\item{target}{Vector of data where some values are missing}

\item{extract_name}{Function for naming discretized features}
}
\description{
This is a generic method that dispatches based on the first argument.
}
