% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma.reports.R
\name{render.Comma}
\alias{render.Comma}
\title{Render the RMD report of comma cases.}
\usage{
render.Comma(data, human.readable)
}
\arguments{
\item{data}{A list of `CommaSummary` objects}

\item{human.readable}{A string to represent the data.}
}
\description{
Render the RMD report of comma cases.
}
