% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{report.DQ}
\alias{report.DQ}
\title{Generate a data quality report for a data set.}
\usage{
report.DQ(
  data,
  human.readable,
  output_pattern = "DQ reports/DQ {human.readable}"
)
}
\arguments{
\item{data}{The data set to be reported ob}

\item{human.readable}{A string representative of the data set to be used
in the report.}

\item{output_pattern}{A string to determine export location.  NB: string is parsed via
`glue` hence code chunks can be included.}
}
\description{
Generate a data quality report for a data set.
}
