% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{report.loc_date}
\alias{report.loc_date}
\title{Generate a missingness report reflecting how dependent missingess is on given
location (categorical) and date time (numeric) variables.}
\usage{
report.loc_date(
  data,
  human.readable,
  location.var,
  date.var,
  output_pattern = "Location Date reports/LD {human.readable}"
)
}
\arguments{
\item{data}{The data set to be reported ob}

\item{human.readable}{A string representative of the data set to be used
in the report.}

\item{location.var}{A variable of `data` that represents a categorical location.}

\item{date.var}{A variable of `data` that represents a numeric date time.}

\item{output_pattern}{A string to determine export location.  NB: string is parsed via
`glue` hence code chunks can be included.}
}
\description{
Generate a missingness report reflecting how dependent missingess is on given
location (categorical) and date time (numeric) variables.
}
