% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{report.missing}
\alias{report.missing}
\title{Generate a general report of missingness for a data set.}
\usage{
report.missing(
  data,
  human.readable,
  output_pattern = "missing_Reports/Missing {human.readable}.html"
)
}
\arguments{
\item{data}{The data set to be reported ob}

\item{human.readable}{A string representative of the data set to be used
in the report.}

\item{output_pattern}{A string to determine export location.  NB: string is parsed via
`glue` hence code chunks can be included.}
}
\description{
Generate a general report of missingness for a data set.
}
