% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma.reports.R
\name{report}
\alias{report}
\title{Check all available sql tables that match a string pattern.}
\usage{
report(con, pattern)
}
\arguments{
\item{con}{A `DBIConnection` object}

\item{pattern}{A string for matching to (grepl)}
}
\description{
Check all available sql tables that match a string pattern.
}
