% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma.reports.R
\name{select_commas}
\alias{select_commas}
\title{Function to filter a list of `CommaSummary` objects for those where
some commas are used.}
\usage{
select_commas(report, limit = 0)
}
\arguments{
\item{report}{list of `CommaSummary` objects}

\item{limit}{[optional] Count of minimum number of commas allowed}
}
\description{
Function to filter a list of `CommaSummary` objects for those where
some commas are used.
}
