% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma.reports.R
\name{zero.prop}
\alias{zero.prop}
\title{Calculate what proportion of responses had zero commas.}
\usage{
zero.prop(comma.summary)
}
\arguments{
\item{comma.summary}{A list produced by `check.last`}
}
\description{
Calculate what proportion of responses had zero commas.
}
